gemini/requirements.txt
flask>=3.0.0,<4
flask-cors>=4.0.0,<5
google-generativeai>=0.7.0
python-dotenv>=1.0.1

